requirements.txt
aiogram>=3.0.0
fastapi
uvicorn
sqlalchemy
pydantic-settings
jinja2
python-multipart
python-dotenv
aiohttp
aioschedule
aiosqlite
pymupdf
